Move parsing methods to FunctionParser
